Update type from number to integer in models and documentation, and bump package version to 1.0.0.6

+++ chatwoot_client/apis/contact_api.py
@@ -44,7 +44,7 @@ class ContactApi(object):
         :param async_req bool
         :param IdContactInboxesBody body: (required)
         :param int account_id: The numeric ID of the account (required)
-        :param float id: ID of the contact (required)
+        :param int id: ID of the contact (required)
         :return: ContactInboxes
                  If the method is called asynchronously,
                  returns the request thread.
@@ -68,7 +68,7 @@ class ContactApi(object):
         :param async_req bool
         :param IdContactInboxesBody body: (required)
         :param int account_id: The numeric ID of the account (required)
-        :param float id: ID of the contact (required)
+        :param int id: ID of the contact (required)
         :return: ContactInboxes
                  If the method is called asynchronously,
                  returns the request thread.
@@ -158,7 +158,7 @@ class ContactApi(object):
 
         :param async_req bool
         :param int account_id: The numeric ID of the account (required)
-        :param float id: ID of the contact (required)
+        :param int id: ID of the contact (required)
         :return: ContactableInboxes
                  If the method is called asynchronously,
                  returns the request thread.
@@ -181,7 +181,7 @@ class ContactApi(object):
 
         :param async_req bool
         :param int account_id: The numeric ID of the account (required)
-        :param float id: ID of the contact (required)
+        :param int id: ID of the contact (required)
         :return: ContactableInboxes
                  If the method is called asynchronously,
                  returns the request thread.
