+++ chatwoot_client/models/agent_bot.py
@@ -28,10 +28,10 @@ class AgentBot(object):
                             and the value is json key in definition.
     """
     swagger_types = {
-        'id': 'float',
+        'id': 'int',
         'name': 'str',
         'description': 'str',
-        'account_id': 'float',
+        'account_id': 'int',
         'outgoing_url': 'str'
     }
 
@@ -69,7 +69,7 @@ class AgentBot(object):
         ID of the agent bot  # noqa: E501
 
         :return: The id of this AgentBot.  # noqa: E501
-        :rtype: float
+        :rtype: int
         """
         return self._id
 
@@ -80,7 +80,7 @@ class AgentBot(object):
         ID of the agent bot  # noqa: E501
 
         :param id: The id of this AgentBot.  # noqa: E501
-        :type: float
+        :type: int
         """
 
         self._id = id
@@ -138,7 +138,7 @@ class AgentBot(object):
         Account ID if it's an account specific bot  # noqa: E501
 
         :return: The account_id of this AgentBot.  # noqa: E501
-        :rtype: float
+        :rtype: int
         """
         return self._account_id
 
@@ -149,7 +149,7 @@ class AgentBot(object):
         Account ID if it's an account specific bot  # noqa: E501
 
         :param account_id: The account_id of this AgentBot.  # noqa: E501
-        :type: float
+        :type: int
         """
 
         self._account_id = account_id
